+++ Data/Domain/OrderDomain.h
@@ -25,7 +25,7 @@ class OrderDomain {
     int number;
     std::tm date;
     OrderStatus status;
-    std::vector<std::pair<Product, int>> products;
+    std::vector<std::pair<Product, float>> products;
     Customer customer;
     employeedomain::Employee employee;
     double totalprice;
@@ -37,7 +37,7 @@ class OrderDomain {
     int getNumber() const;
     std::tm getDate() const;
     OrderStatus getStatus() const;
-    std::vector<std::pair<Product, int>> getProducts() const;
+    std::vector<std::pair<Product, float>> getProducts() const;
     double getTotalPrice() const;
     employeedomain::Employee getEmployee() const;
     Customer getCustomer() const;
@@ -50,7 +50,7 @@ class OrderDomain {
     void setEmployee(const employeedomain::Employee& newemployee);
 
     //setter method for changing the list of products
-    void setProducts(const std::vector<std::pair<Product, int>>& newproducts);
+    void setProducts(const std::vector<std::pair<Product, float>>& newproducts);
 
 
     //methods for employee
@@ -58,13 +58,13 @@ class OrderDomain {
     void changeStatus(OrderStatus newStatus);
 
     //changes the list of products to a new list
-    void changeOrder(const std::vector<std::pair<Product, int>>& newProducts);
+    void changeOrder(const std::vector<std::pair<Product, float>>& newProducts);
 
     //calculates the total price
     void calculateTotalPrice();
 
     //creates a "Reservation" order, that doesn't have an assigned employee
-    void createReservation(const Customer& customer, const std::vector<std::pair<Product, int>>& products);
+    void createReservation(const Customer& customer, const std::vector<std::pair<Product, float>>& products);
 
 
 };
